backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
